[Canvas] `SidebarHeader` refactor. (#110176)
* Removed recompose.
