+++ x-pack/plugins/canvas/public/components/sidebar_header/index.tsx
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+export { SidebarHeader } from './sidebar_header';
+export { SidebarHeader as SidebarHeaderComponent } from './sidebar_header.component';
